refactor: move all validation code to validation.js
