+++ src/routes/auth.js
@@ -2,12 +2,13 @@ const express = require('express');
 const bcrypt = require('bcrypt');
 const Joi = require('joi');
 const router = express.Router();
+const {validateLogin} = require('../middleware/validation');
 
 const { User } = require('../models/user');
 const { Consultant } = require('../models/consultant');
 
 router.post('/user', async (req, res) => {
-  const { error } = validate(req.body);
+  const { error } = validateLogin(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
   let user = await User.findOne({ email: req.body.email });
@@ -21,7 +22,7 @@ router.post('/user', async (req, res) => {
 });
 
 router.post('/consultant', async (req, res) => {
-  const { error } = validate(req.body);
+  const { error } = validateLogin(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
   let consultant = await Consultant.findOne({ email: req.body.email });
@@ -33,14 +34,5 @@ router.post('/consultant', async (req, res) => {
   const token = consultant.generateAuthToken();
   res.send(token);
 });
-
-function validate(req) {
-  const schema = {
-    email: Joi.string().max(255).required().email(),
-    password: Joi.string().min(6).max(255).required().strict()
-  };
-
-  return Joi.validate(req, schema);
-}
 
 module.exports = router;
