+++ src/routes/consultants.js
@@ -3,7 +3,8 @@ const router = express.Router();
 const bcrypt = require('bcrypt');
 const _ = require('lodash');
 const mongoose = require('mongoose');
-const { Consultant, validateConsultant } = require('../models/consultant');
+const { Consultant } = require('../models/consultant');
+const { validateConsultant } = require('../middleware/validation');
 
 router.post('/', async (req, res) => {
 
@@ -16,7 +17,7 @@ router.post('/', async (req, res) => {
   if (consultant) return res.status(400).send('consultant already registered');
 
   try {
-    consultant = new Consultant(_.pick(req.body, ['firstname', 'lastname', 'email', 'password', 'phone', 'certification','workplace']));
+    consultant = new Consultant(_.pick(req.body, ['firstname', 'lastname', 'email', 'password', 'phone', 'certification', 'workplace']));
     const salt = await bcrypt.genSalt(10);
 
     //Hashes consultant's password
@@ -25,7 +26,7 @@ router.post('/', async (req, res) => {
 
     //Generates token and returns it as a header for auto auth
     const token = consultant.generateAuthToken();
-    res.header('x-auth-token', token).send(_.pick(consultant, ['firstname', 'lastname', 'email', 'phone', 'certification','workplace']));
+    res.header('x-auth-token', token).send(_.pick(consultant, ['firstname', 'lastname', 'email', 'phone', 'certification', 'workplace']));
 
   } catch (error) {
